Move notes back to actor
Essentially, I'd like every entity (actor, item) to have a notes field;
it's the catch-all for anything where people can write or keep track of
what they need. Lifepath is a great place to store and prompt summaries
of their lives, but there still needs to be a notes field overall, and
players probably wouldn't look in Life for it.

Also changed Style in the localization file to Style & Origin, as that
section includes both (ahhh, cyberpunk)